+++ src/components/modules/popup/details/ColorsUnit.tsx
@@ -1,10 +1,14 @@
 import type {ColorData} from '_scripts/colorsExtractor'
+import {ITEMS_STYLE} from '~/Global/Header'
+import {ROUTES} from '@/lib/routes'
 
 import {useState} from 'react'
 import {cn} from '@/lib/utils'
+import {ArrowUpRight} from 'lucide-react'
 
 import Unit from '~/UI/Unit'
 import {SPAN} from '~/UI/Typography'
+import Button from '~/UI/Button'
 
 export default function ColorsUnit({data}: {data: ColorData[] | undefined}) {
   const [tooltip, setTooltip] = useState('')
@@ -31,6 +35,10 @@ export default function ColorsUnit({data}: {data: ColorData[] | undefined}) {
             <SPAN className={cn('absolute -top-[35px] left-1/2 transform -translate-x-1/2', 'px-1.5 py-1 text-background bg-white rounded-md', 'opacity-0 group-hover:opacity-100 transition-opacity')}>{tooltip || color}</SPAN>
           </div>
         ))}
+
+        <Button to={ROUTES.colors} className={cn('px-1', 'grid place-items-center group hover:bg-unit')}>
+          <ArrowUpRight className={cn(ITEMS_STYLE.icon, 'size-10')} strokeWidth={1.25} />
+        </Button>
       </div>
     </Unit>
   )